% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{FizzBuzz}
\alias{FizzBuzz}
\title{FizzBuzz}
\usage{
FizzBuzz(n)
}
\arguments{
\item{n}{is the input value that determines the range 1:n}
}
\value{
a vector of characters with "Fizz", "Buzz", "FizzBuzz", or a number
}
\description{
My function takes a positive integer n and outputs a vector of characters.
In the range 1:n, if the number is divisible by 3, the corresponding index in the
output will say "Fizz"; if the number is divisible by 5, the output will be "Buzz".
If it is divisible by both 3 and 5, the output will say "FizzBuzz". Otherwise
it will  just say the number.
}
\examples{
FizzBuzz(10)

}
